requirements.txt
azure-functions==1.20.0
certifi==2024.6.2
charset-normalizer==3.3.2
idna==3.7
requests==2.32.3
slack_sdk==3.29.0
urllib3==2.2.2
